merubah bagian public down di migrationnya dan hapus notif yg lama

+++ app/Database/Migrations/2025-07-01-061852_Document.php
@@ -94,6 +94,8 @@ class Document extends Migration
 
     public function down()
     {
-        $this->forge->dropTable('document');
+        if ($this->db->tableExists('document')) {
+            $this->forge->dropTable('document');
+        }
     }
 }
