+++ app/Database/Migrations/2025-07-01-064213_Privilege.php
@@ -63,6 +63,8 @@ class Privilege extends Migration
 
     public function down()
     {
-        $this->forge->dropTable('privilege');
+        if ($this->db->tableExists('privilege')) {
+            $this->forge->dropTable('privilege');
+        }
     }
 }
