+++ app/Database/Migrations/2025-07-25-014027_UserWC.php
@@ -57,6 +57,8 @@ class UserWC extends Migration
 
     public function down()
     {
-        $this->forge->dropTable('userwc');
+        if ($this->db->tableExists('userwc')) {
+            $this->forge->dropTable('userwc');
+        }
     }
 }
